move EntryActionRepository to :domain:identity:usecase
