+++ domain/identity/repository/src/androidMain/kotlin/com/livefast/eattrash/raccoonforfriendica/domain/identity/repository/di/Utils.kt
@@ -2,7 +2,6 @@ package com.livefast.eattrash.raccoonforfriendica.domain.identity.repository.di
 
 import com.livefast.eattrash.raccoonforfriendica.domain.identity.repository.ApiConfigurationRepository
 import com.livefast.eattrash.raccoonforfriendica.domain.identity.repository.AuthManager
-import com.livefast.eattrash.raccoonforfriendica.domain.identity.repository.EntryActionRepository
 import com.livefast.eattrash.raccoonforfriendica.domain.identity.repository.SettingsRepository
 import org.koin.java.KoinJavaComponent
 
@@ -18,10 +17,5 @@ actual fun getSettingsRepository(): SettingsRepository {
 
 actual fun getAuthManager(): AuthManager {
     val res by KoinJavaComponent.inject<AuthManager>(AuthManager::class.java)
-    return res
-}
-
-actual fun getEntryActionRepository(): EntryActionRepository {
-    val res by KoinJavaComponent.inject<EntryActionRepository>(EntryActionRepository::class.java)
     return res
 }
