+++ domain/identity/repository/src/commonMain/kotlin/com/livefast/eattrash/raccoonforfriendica/domain/identity/repository/di/Utils.kt
@@ -2,7 +2,6 @@ package com.livefast.eattrash.raccoonforfriendica.domain.identity.repository.di
 
 import com.livefast.eattrash.raccoonforfriendica.domain.identity.repository.ApiConfigurationRepository
 import com.livefast.eattrash.raccoonforfriendica.domain.identity.repository.AuthManager
-import com.livefast.eattrash.raccoonforfriendica.domain.identity.repository.EntryActionRepository
 import com.livefast.eattrash.raccoonforfriendica.domain.identity.repository.SettingsRepository
 
 expect fun getApiConfigurationRepository(): ApiConfigurationRepository
@@ -10,5 +9,3 @@ expect fun getApiConfigurationRepository(): ApiConfigurationRepository
 expect fun getSettingsRepository(): SettingsRepository
 
 expect fun getAuthManager(): AuthManager
-
-expect fun getEntryActionRepository(): EntryActionRepository
