+++ domain/identity/usecase/src/commonMain/kotlin/com/livefast/eattrash/raccoonforfriendica/domain/identity/usecase/DefaultEntryActionRepository.kt
@@ -1,9 +1,12 @@
-package com.livefast.eattrash.raccoonforfriendica.domain.identity.repository
+package com.livefast.eattrash.raccoonforfriendica.domain.identity.usecase
 
 import com.livefast.eattrash.raccoonforfriendica.domain.content.data.TimelineEntryModel
+import com.livefast.eattrash.raccoonforfriendica.domain.content.repository.SupportedFeatureRepository
+import com.livefast.eattrash.raccoonforfriendica.domain.identity.repository.IdentityRepository
 
 internal class DefaultEntryActionRepository(
     private val identityRepository: IdentityRepository,
+    private val supportedFeatureRepository: SupportedFeatureRepository,
 ) : EntryActionRepository {
     private val currentUserId: String? get() = identityRepository.currentUser.value?.id
     private val isLogged: Boolean get() = currentUserId != null
@@ -29,6 +32,8 @@ internal class DefaultEntryActionRepository(
     override fun canTogglePin(entry: TimelineEntryModel): Boolean = entry.isFromCurrentUser
 
     override fun canBlock(entry: TimelineEntryModel): Boolean = entry.isFromOtherUser
+
+    override fun canQuote(entry: TimelineEntryModel): Boolean = supportedFeatureRepository.features.value.supportsEntryShare
 
     private val TimelineEntryModel.isFromCurrentUser: Boolean
         get() = isLogged && creator?.id == currentUserId
