+++ domain/identity/usecase/src/commonMain/kotlin/com/livefast/eattrash/raccoonforfriendica/domain/identity/usecase/di/Utils.kt
@@ -3,6 +3,7 @@ package com.livefast.eattrash.raccoonforfriendica.domain.identity.usecase.di
 import androidx.compose.ui.platform.UriHandler
 import com.livefast.eattrash.raccoonforfriendica.domain.identity.usecase.ActiveAccountMonitor
 import com.livefast.eattrash.raccoonforfriendica.domain.identity.usecase.CustomUriHandler
+import com.livefast.eattrash.raccoonforfriendica.domain.identity.usecase.EntryActionRepository
 import com.livefast.eattrash.raccoonforfriendica.domain.identity.usecase.SetupAccountUseCase
 
 expect fun getActiveAccountMonitor(): ActiveAccountMonitor
@@ -10,3 +11,5 @@ expect fun getActiveAccountMonitor(): ActiveAccountMonitor
 expect fun getSetupAccountUseCase(): SetupAccountUseCase
 
 expect fun getCustomUriHandler(fallback: UriHandler): CustomUriHandler
+
+expect fun getEntryActionRepository(): EntryActionRepository
